requirements.txt
Django==4.1.7
djangorestframework==3.13.1
django-cors-headers==3.11.0
mysqlclient==2.2.4
pycryptodome
python-alipay-sdk